Fix the priority crash
The `bridgifyObject` call is not working correctly, since we are using a different YAOB instance than the core.

Instead of dealing with this, we we send across the raw C++ API. Then we hijack the CppBridge object and instantiate it on the WebView side, using the C++ API we passed earlier. Yes, this is a hack.