Don't update tx when native amounts change
The tx objects returned from the SDK don't always include both the received and sent amounts for mempool and freshly mined transactions. This causes the amount to change back and forth between a couple values before it settles on the value we saved initially. The rest of the edgeTransaction is fine but we can trust the amounts we already had.