Remove separate ESM build output
Neither Metro nor Node.js take advantage of this, so don't bother including it.

This is not a breaking change, because any packager that used the non-standard `module` field would also know how to handle the `main` field. The only side-effect would be slightly larger builds output, but not breakage.